+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+
+* Add `Pool.close()`, which forbids new resource requests and releases all
+  releasable resources.
+
 ## 1.1.0
 
 * Add `PoolResource.allowRelease()`, which allows a resource to indicate that it